feat: anadir listar albunes y arquitectura inicial project, actualizar a java 17
