+++ app/src/main/java/com/uniandes/vinilos/model/Album.kt
@@ -5,21 +5,10 @@ data class Album(
     val name: String,
     val cover: String,
     val releaseDate: String,
+    val description: String,
     val genre: String,
     val recordLabel: String,
-    val description: String,
     val tracks: List<Track> = emptyList(),
-    val performers: List<Performer> = emptyList()
-)
-
-data class Track(
-    val id: Int,
-    val name: String,
-    val duration: String
-)
-
-data class Performer(
-    val id: Int,
-    val name: String,
-    val image: String
-)
+    val performers: List<Performer> = emptyList(),
+    val comments: List<Comment> = emptyList()
+)
