+++ app/src/main/java/com/uniandes/vinilos/model/Comment.kt
@@ -1,0 +1,7 @@
+package com.uniandes.vinilos.model
+
+data class Comment(
+    val id: Int,
+    val description: String,
+    val rating: Int
+)
