+++ app/src/main/java/com/uniandes/vinilos/model/Performer.kt
@@ -1,0 +1,10 @@
+package com.uniandes.vinilos.model
+
+data class Performer(
+    val id: Int,
+    val name: String,
+    val image: String,
+    val description: String,
+    val birthDate: String? = null,
+    val creationDate: String? = null
+)
