+++ app/src/main/java/com/uniandes/vinilos/model/Track.kt
@@ -1,0 +1,7 @@
+package com.uniandes.vinilos.model
+
+data class Track(
+    val id: Int,
+    val name: String,
+    val duration: String
+)
