+++ app/src/main/java/com/uniandes/vinilos/network/AlbumService.kt
@@ -3,12 +3,11 @@ package com.uniandes.vinilos.network
 import com.uniandes.vinilos.model.Album
 import retrofit2.http.GET
 import retrofit2.http.Path
-import retrofit2.http.Url
 
 interface AlbumService {
-    @GET
-    suspend fun getAlbums(@Url url: String = "${ApiConfig.ALBUMS_BASE_URL}"): List<Album>
+    @GET("albums")
+    suspend fun getAlbums(): List<Album>
 
-    @GET
-    suspend fun getAlbum(@Url url: String = "${ApiConfig.ALBUMS_BASE_URL}/{id}"): Album
+    @GET("albums/{id}")
+    suspend fun getAlbum(@Path("id") id: Int): Album
 }
